make conn as tx
